backend/requirements.txt
Django>=5.0
djangorestframework
djangorestframework-simplejwt
mysqlclient
python-dotenv
django-cors-headers
Pillow
dj-database-url
psycopg2-binary
gunicorn
whitenoise
